+++ packages/patternfly-4/react-core/src/components/Accordion/Accordion.d.ts
@@ -1,0 +1,10 @@
+import { FunctionComponent, HTMLProps } from 'react';
+import { Omit } from '../../typeUtils';
+
+export interface AccordionProps extends Omit<HTMLProps<HTMLUListElement>, 'aria-label'> {
+  'aria-label': string;
+}
+
+declare const Accordion: FunctionComponent<AccordionProps>;
+
+export default Accordion;