+++ packages/patternfly-4/react-core/src/components/Accordion/AccordionItem.d.ts
@@ -1,0 +1,9 @@
+import { FunctionComponent, HTMLProps } from 'react';
+import { Omit } from '../../typeUtils';
+
+export interface AccordionItemProps extends Omit<HTMLProps<HTMLLIElement>, 'aria-label'> {
+}
+
+declare const AccordionItem: FunctionComponent<AccordionItemProps>;
+
+export default AccordionItem;