+++ packages/patternfly-4/react-core/src/components/Accordion/AccordionToggle.d.ts
@@ -1,0 +1,14 @@
+import { FunctionComponent, HTMLProps } from 'react';
+import { Omit } from '../../typeUtils';
+
+export interface AccordionToggleProps extends Omit<HTMLProps<HTMLDivElement>, 'aria-labelledby' | 'aria-label' | 'id'> {
+  isExpanded: boolean;
+  'aria-controls'?: string;
+  'aria-labelledby'?: string;
+  'aria-label'?: string;
+  id: string;
+}
+
+declare const AccordionToggle: FunctionComponent<AccordionToggleProps>;
+
+export default AccordionToggle;