+++ packages/patternfly-4/react-core/src/components/Accordion/index.js
@@ -1,0 +1,4 @@
+export { default as Accordion } from './Accordion';
+export { default as AccordionItem } from './AccordionItem';
+export { default as AccordionContent } from './AccordionContent';
+export { default as AccordionToggle } from './AccordionToggle';