+++ packages/patternfly-4/react-core/src/components/Checkbox/Checkbox.test.tsx
@@ -1,6 +1,6 @@
-import React from 'react';
+import * as React from 'react';
 import { shallow } from 'enzyme';
-import Checkbox from './Checkbox';
+import { Checkbox } from './Checkbox';
 
 const props = {
   onChange: jest.fn(),