+++ packages/patternfly-4/react-core/src/components/Dropdown/DropdownToggleCheckbox.d.ts
@@ -1,16 +1,15 @@
 import { FormEvent, FunctionComponent, HTMLProps, ReactNode } from 'react';
 import { Omit } from '../../helpers/typeUtils';
 
-export interface CheckboxProps extends Omit<HTMLProps<HTMLInputElement>, 'type' | 'onChange' | 'disabled' | 'label'> {
+export interface DropdownToggleCheckboxProps extends Omit<HTMLProps<HTMLInputElement>, 'type' | 'onChange' | 'disabled'> {
   isDisabled?: boolean;
   isValid?: boolean;
   isChecked?: boolean;
   onChange?(checked: boolean, event: FormEvent<HTMLInputElement>): void;
   id: string;
   'aria-label': string;
-  label?: ReactNode;
 }
 
-declare const Checkbox: FunctionComponent<CheckboxProps>;
+declare const DropdownToggleCheckbox: FunctionComponent<DropdownToggleCheckboxProps>;
 
-export default Checkbox;
+export default DropdownToggleCheckbox;