+++ packages/patternfly-4/react-core/src/components/Select/SelectOption.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import styles from '@patternfly/patternfly/components/Select/select.css';
 import { css } from '@patternfly/react-styles';
+import { CheckIcon } from '@patternfly/react-icons';
 import PropTypes from 'prop-types';
 import { SelectContext, KeyTypes } from './selectConstants';
 
@@ -79,7 +80,7 @@ class SelectOption extends React.Component {
               {...props}
               className={css(
                 styles.selectMenuItem,
-                isSelected && styles.selectMenuItemMatch,
+                isSelected && styles.modifiers.selected,
                 isDisabled && styles.modifiers.disabled,
                 className
               )}
@@ -96,6 +97,7 @@ class SelectOption extends React.Component {
               onKeyDown={this.onKeyDown}
             >
               {value}
+              {isSelected && <CheckIcon noVerticalAlign className={css(styles.selectMenuItemIcon)} aria-hidden />}
             </button>
           </li>
         )}