+++ packages/patternfly-4/react-core/src/components/index.ts
@@ -1,5 +1,6 @@
 /** Keep alphabetically sorted */
 export * from './AboutModal';
+// export * from './Accordion'; Don't release this yet
 export * from './Alert';
 export * from './ApplicationLauncher';
 export * from './Avatar';
@@ -26,12 +27,14 @@ export * from './List';
 export * from './LoginPage';
 export * from './Modal';
 export * from './Nav';
+export * from './OptionsMenu';
 export * from './Page';
 export * from './Pagination';
 export * from './Popover';
 export * from './Progress';
 export * from './Radio';
 export * from './Select';
+export * from './SkipToContent';
 export * from './Switch';
 export * from './Tabs';
 export * from './Text';
@@ -39,3 +42,4 @@ export * from './TextArea';
 export * from './TextInput';
 export * from './Title';
 export * from './Tooltip';
+export * from './Wizard';